enclosure service created
