+++ src/app/model/enclosure.model.ts
@@ -1,0 +1,5 @@
+export interface TemperatureReading {
+  temperature: number;
+  humidity: number;
+  unit: string;
+}
