+++ src/app/plugins/index.ts
@@ -1,7 +1,6 @@
 export * from '../services/filament/filament.service';
 export * from '../services/filament/filament-manager.octoprint.service';
 export * from '../services/filament/filament-plugin.service';
-export * from './enclosure.service';
 export * from './layer-progress.service';
 export * from './psu-control.service';
 export * from './tplink-smartplug.service';
