+++ src/app/services/enclosure/enclosure.service.ts
@@ -1,0 +1,19 @@
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+
+import { TemperatureReading } from '../../model/enclosure.model';
+
+// TODO
+
+// PsuControlService
+// TPLink SmartPlug
+// Enclosure IO
+
+@Injectable()
+export abstract class EnclosureService {
+  abstract getEnclosureTemperature(): Observable<TemperatureReading>;
+
+  abstract setLEDColor(identifier: number, red: number, green: number, blue: number): void;
+
+  abstract setOutput(identifier: number, status: boolean): void;
+}
